:lipstick: feat: mudando interaface de task completa

+++ front-tarefas/src/Services/Tasks.js
@@ -2,8 +2,12 @@ import axios from 'axios';
 
 const API_URL = `${import.meta.env.VITE_API_URL}/tasks`;
 
-export const getTasks = async (token, date) => {
+export const getTasks = async (date) => {
     try {
+        const token = localStorage.getItem('token'); // Certifique-se de obter o token aqui
+        if (!token) {
+            throw new Error('Token não encontrado');
+        }
         const response = await axios.get(API_URL, {
             headers: {
                 'Authorization': `Bearer ${token}`
@@ -15,6 +19,7 @@ export const getTasks = async (token, date) => {
         console.log("Resposta da API no getTasks:", response.data);
         return response.data;
     } catch (error) {
+        console.error('Erro ao buscar tarefas:', error); // Adicione logs para depuração
         throw new Error('Erro ao buscar tarefas');
     }
 };
@@ -22,6 +27,9 @@ export const getTasks = async (token, date) => {
 export const createTask = async (task) => {
     try {
         const token = localStorage.getItem('token');
+        if (!token) {
+            throw new Error('Token não encontrado');
+        }
         const response = await axios.post(API_URL, task, {
             headers: {
                 'Content-Type': 'application/json',
@@ -31,6 +39,7 @@ export const createTask = async (task) => {
         console.log("Resposta da API no createTask:", response.data);
         return response.data;
     } catch (error) {
+        console.error('Erro ao criar tarefa:', error); // Adicione logs para depuração
         throw new Error('Erro ao criar tarefa');
     }
 };
@@ -38,6 +47,9 @@ export const createTask = async (task) => {
 export const updateTask = async (taskId, task) => {
     try {
         const token = localStorage.getItem('token');
+        if (!token) {
+            throw new Error('Token não encontrado');
+        }
         const response = await axios.put(`${API_URL}/${taskId}`, task, {
             headers: {
                 'Content-Type': 'application/json',
@@ -47,6 +59,7 @@ export const updateTask = async (taskId, task) => {
         console.log("Resposta da API no updateTask:", response.data);
         return response.data;
     } catch (error) {
+        console.error('Erro ao atualizar tarefa:', error); // Adicione logs para depuração
         throw new Error('Erro ao atualizar tarefa');
     }
 };
@@ -54,6 +67,9 @@ export const updateTask = async (taskId, task) => {
 export const deleteTask = async (taskId) => {
     try {
         const token = localStorage.getItem('token');
+        if (!token) {
+            throw new Error('Token não encontrado');
+        }
         const response = await axios.delete(`${API_URL}/${taskId}`, {
             headers: {
                 'Authorization': `Bearer ${token}`
@@ -62,6 +78,7 @@ export const deleteTask = async (taskId) => {
         console.log("Resposta da API no deleteTask:", response.data);
         return response.data;
     } catch (error) {
+        console.error('Erro ao deletar tarefa:', error); // Adicione logs para depuração
         throw new Error('Erro ao deletar tarefa');
     }
 };
@@ -69,6 +86,9 @@ export const deleteTask = async (taskId) => {
 export const toggleCompleteTask = async (taskId) => {
     try {
         const token = localStorage.getItem('token');
+        if (!token) {
+            throw new Error('Token não encontrado');
+        }
         const response = await axios.patch(`${API_URL}/${taskId}/toggle-complete`, {}, {
             headers: {
                 'Content-Type': 'application/json',
@@ -78,6 +98,7 @@ export const toggleCompleteTask = async (taskId) => {
         console.log("Resposta da API no toggleCompleteTask:", response.data);
         return response.data;
     } catch (error) {
+        console.error('Erro ao atualizar tarefa:', error); // Adicione logs para depuração
         throw new Error('Erro ao atualizar tarefa');
     }
 };
